apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.19
pydantic-settings==2.3.1
anthropic==0.28.0
httpx==0.27.0
feedparser==6.0.11
python-frontmatter==1.1.0
apscheduler==3.10.4
watchdog==4.0.1
python-multipart==0.0.9
